+++ cypress/e2e/agreements/agreement-create-and-delete.cy.js
@@ -7,6 +7,7 @@ import AppInteractor from '../../support/fragments/agreements/AppInteractor';
 import AgreementViewInteractor from '../../support/fragments/agreements/AgreementViewInteractor';
 
 import generateItemBarcode from '../../support/utils/generateItemBarcode';
+import AgreementsSettingsInteractor from '../../support/fragments/agreements/AgreementsSettingsInteractor';
 
 const agreementName = 'Test: ' + generateItemBarcode();
 
@@ -45,7 +46,7 @@ describe('Agreement create and delete', () => {
     cy.createUserWithPwAndPerms(viewUser, viewPermissions);
 
     cy.getAdminToken();
-    AppInteractor.fetchStatusLabel(agreement);
+    AgreementsSettingsInteractor.fetchStatusLabel(agreement);
   });
 
   after(() => {
@@ -86,10 +87,7 @@ describe('Agreement create and delete', () => {
 
       // it has to be 'function ()' and NOT '() =>' because otherwise the alias can't be accessed with this.*
       it('should open "Record last updated" information and see correct values', function () {
-        cy.get('[id=agreementInfoRecordMeta]').click().within(() => {
-          cy.contains('Record created: ' + DateTools.getFormattedDateWithTime(this.dateCreated));
-          cy.contains('Record last updated: ' + DateTools.getFormattedDateWithTime(this.dateCreated));
-        });
+        AgreementViewInteractor.recordMetadataInfo(this.dateCreated);
       });
     });
   }