+++ cypress/support/fragments/agreements/AppInteractor.js
@@ -10,8 +10,6 @@ import {
   Section,
 } from '@folio/stripes-testing';
 
-import { normalize } from '../../utils/stringTools';
-
 /* We can import other interactors here and expose their functionality
  * to allow for a singular "AppInteractor" import in our tests.
  * I'm not sure if that's a good idea or not, I quite like the idea that different
@@ -98,16 +96,4 @@ export default class AppInteractor {
       Button('Search').click()
     ]);
   };
-
-  static fetchStatusLabel = (agreement) => {
-    const refdataDesc = 'SubscriptionAgreement.AgreementStatus';
-    cy.getAgreementsRefdataValues(refdataDesc).then((refdata) => {
-      if (refdata.every(obj => obj.label !== agreement.status)) {
-        cy.getAgreementsRefdataLabelFromValue(refdataDesc, normalize(agreement.status))
-          .then((refdataLabel) => {
-            agreement.status = refdataLabel;
-          });
-      }
-    });
-  }
 }