+++ cypress/support/fragments/licenses/AppInteractor.js
@@ -1,5 +1,6 @@
 import {
   Button,
+  DropdownMenu,
   HTML,
   including,
   MultiColumnListCell,
@@ -10,8 +11,6 @@ import {
   Section,
 } from '@folio/stripes-testing';
 
-import { normalize } from '../../utils/stringTools';
-
 /* We can import other interactors here and expose their functionality
  * to allow for a singular "AppInteractor" import in our tests.
  * I'm not sure if that's a good idea or not, I quite like the idea that different
@@ -19,20 +18,24 @@ import { normalize } from '../../utils/stringTools';
  * such as creating a new license, to be controlled via the AppInteractor does make some sense.
  */
 import LicenseFormInteractor from './LicenseFormInteractor';
+import LicenseViewInteractor from './LicenseViewInteractor';
+import HomeInteractor from '../HomeInteractor';
 
 // Making Appinteractor a class so it is consistent and clear where each action is coming from
 // The main interactor for the licenses 3 pane main landing page
 export default class AppInteractor {
   static section = Section({ id: 'pane-license-list' });
 
-  static actionButton = Button('Actions');
+  static actionsButton = Pane(including('Licenses')).find(Button('Actions'));
+  static newButton = DropdownMenu().find(Button('New'));
+  static exportCSVButton = DropdownMenu().find(Button({ id: 'export-licenses-csv' }));
 
   static waitLoading = () => {
     cy.expect(or(
       this.section.find(MultiColumnListRow()).exists(),
       this.section.find(HTML(including('No results found. Please check your filters.'))).exists()
     ));
-    cy.expect(this.actionButton.exists());
+    cy.expect(this.actionsButton.exists());
   };
 
   static createLicense = (license) => {
@@ -54,6 +57,15 @@ export default class AppInteractor {
     ]);
   };
 
+  /*
+  unlike in agreements app
+  license is not selected automatically after a search with single result
+  */
+  static selectLicense = (licenseName) => {
+    cy.do(this.section.find(MultiColumnListCell(licenseName)).click());
+    LicenseViewInteractor.paneExists(licenseName);
+  }
+
   static licenseNotVisible = (licenseTitle) => {
     cy.expect(or(
       this.section.find(MultiColumnListCell(licenseTitle)).absent(),
@@ -61,15 +73,13 @@ export default class AppInteractor {
     ));
   };
 
-  static fetchStatusLabel = (license) => {
-    const refdataDesc = 'License.Status';
-    cy.getLicensesRefdataValues(refdataDesc).then((refdata) => {
-      if (refdata.every(obj => obj.label !== license.status)) {
-        cy.getLicensesRefdataLabelFromValue(refdataDesc, normalize(license.status))
-          .then((refdataLabel) => {
-            license.status = refdataLabel;
-          });
-      }
-    });
-  }
+  static licenseVisible = (licenseTitle) => {
+    cy.expect(this.section.find(MultiColumnListCell(licenseTitle)).exists());
+  };
+
+  static openLicensesApp = () => {
+    HomeInteractor.navToApp('Licenses');
+    cy.url().should('include', '/licenses');
+    cy.expect(Pane(including('Licenses')).exists());
+  };
 }