+++ CHANGELOG.md
@@ -4,7 +4,11 @@
   * Added stripes-kint-components to dev deps
   * Tweaked Harnesses so that we can export *just* the base harness without the kint intl stuff. This means that kint-components can use the base harness as well.
   * ERM-2773 Add agreement line with internal KB resource
-    * add packages: uuid, cypress-wait-until 
+    * add packages: uuid, cypress-wait-until
+  * ERM-2812 Create an Agreement and Delete an Agreement
+    * extend select interactor with selectedContent filter
+  * ERM-2715 Create a license and delete a license
+    * add a bunch of new interactor functions and cypress commands
 
 ## 1.1.0 2023-02-22
   * Initial e2e cypress tests